requirements.txt
cat > requirements.txt << 'EOF'
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
openai==1.3.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
requests==2.31.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
EOF